improve tests by not starting gateway on each test.
